Move class member related verification code to Translator.ClassMembers.cs
